Fix Makefile to use libsilk from current directory add log file for echo_client & echo_server msg len is uint16_t in header & int on stack. this sometimes caused trouble so explicit casting is now in place & identical type used on stack

+++ src/echo_server.c
@@ -221,7 +221,7 @@ echo_server_entry_func (void *_arg)
         if (echo_server_rcv_buf(conn, &net_msg.hdr, sizeof(net_msg.hdr)) < 0) {
             goto socket_close;
         }
-        printf("msg hdr completed. len=%d\n", net_msg.hdr.len);
+        printf("msg hdr completed. len=%u\n", (unsigned)net_msg.hdr.len);
 
         printf("Reading the msg text...\n");
         if (echo_server_rcv_buf(conn, net_msg.msg, net_msg.hdr.len) < 0) {
